requirements.txt
streamlit
python-dotenv
PyPDF2
pandas
huggingface_hub